requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
filelock==3.8.0
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
murmurhash==1.0.9
nltk==3.7
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.2
pathy==0.10.0
Pillow==9.3.0
preshed==3.0.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
smart-open==5.2.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
thinc==8.1.5
tokenizers==0.13.2
torch==1.12.0
tqdm==4.64.1
transformers==4.24.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
